Schneider building SBO
Schneider building control SBO-485 protocol data docking
According to the existing data table, compare the self-defined digital output input and analog output input to build the TCP server service. Set up the data point binding operation, and define the data point ID according to the type of data. Generally, the data point of digital type is 1 byte, and the data point bit of analog type is 2 bytes.
Finally, the data type is defined by the output and input requirements. This requires us to obtain the Modbus control and controlled data points of the SBO building control system connected to the third-party panel and some 485 protocols. Of course, the logic control can also be done in the third software. It is important to note that the control response time is best placed in the local synchronous data so as to avoid delay. If you have any questions, please send me a message!!!!